move Script e2e test to async-config
