+++ packages/playground/async-config/package.json
@@ -16,6 +16,7 @@
     "@shopify/cli-hydrogen": "^3.12.0"
   },
   "dependencies": {
+    "@builder.io/partytown": "^0.6.4",
     "@cloudflare/kv-asset-handler": "*",
     "@shopify/hydrogen": "^1.4.3",
     "miniflare": "^1.3.3",
