+++ packages/playground/async-config/public/after-hydration-script.js
