+++ packages/playground/async-config/public/before-hydration-script.js
