+++ packages/playground/async-config/public/in-worker-script.js
