+++ packages/playground/async-config/public/load-script.js
