+++ packages/playground/async-config/public/on-idle-script.js
