+++ packages/playground/async-config/public/test-module-script.js
