+++ packages/playground/async-config/src/components/ScriptUseLoadScript.client.jsx
