+++ packages/playground/async-config/src/routes/scripts/cdn.server.jsx
