+++ packages/playground/async-config/src/routes/scripts/index.server.jsx
