+++ packages/playground/async-config/src/routes/scripts/script/after-hydration/[handle].server.jsx
