+++ packages/playground/async-config/src/routes/scripts/script/after-hydration/index.server.jsx
