+++ packages/playground/async-config/src/routes/scripts/script/in-worker/index.server.jsx
