+++ packages/playground/async-config/src/routes/scripts/script/index.server.jsx
