+++ packages/playground/async-config/src/routes/scripts/script/on-idle/[handle].server.jsx
