+++ packages/playground/async-config/src/routes/scripts/script/on-idle/index.server.jsx
