+++ packages/playground/async-config/src/routes/scripts/use-script/[handle].server.jsx
