+++ packages/playground/server-components/package.json
@@ -10,13 +10,12 @@
     "start:worker": "node ../test-utils/start-worker"
   },
   "devDependencies": {
-    "@shopify/cli": "^3.3.3",
-    "@shopify/cli-hydrogen": "^3.3.3",
     "cross-env": "^7.0.3",
-    "vite": "^2.9.13"
+    "vite": "^2.9.13",
+    "@shopify/cli": "^3.12.0",
+    "@shopify/cli-hydrogen": "^3.12.0"
   },
   "dependencies": {
-    "@builder.io/partytown": "^0.6.4",
     "@cloudflare/kv-asset-handler": "*",
     "@shopify/hydrogen": "^1.4.3",
     "miniflare": "^1.3.3",
